requirements.txt
gspread==5.9.0
oauth2client==4.1.3
PyYAML==6.0
schedule==1.2.0
